sike and sidh with interfaces
