+++ sidh/sidhp751/sidh_amd64.s
